wip ias
How to link a deployed enclave with an attestation report? If one has
verified attestation report for an enclave, how does one verify that
they are communicating with the verified enclave? If the enclave signs
outgoing messages with a private key for which the corresponding public
component is included in the report data of the report, then one can
verify the signature, with the public key. This assumes that the private
key can be kept secure such that nothing except the verify enclave can
access it. See key gen and seal.